Errors module renamed as exceptions
